code-mode: extract protocol and add host crate
