+++ codex-rs/code-mode-host/src/main.rs
@@ -1,0 +1,1 @@
+fn main() {}
