+++ codex-rs/code-mode-protocol/BUILD.bazel
@@ -1,0 +1,6 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "code-mode-protocol",
+    crate_name = "codex_code_mode_protocol",
+)
