+++ codex-rs/code-mode-protocol/src/response.rs
